Working on API, CI/CD, Tests
