+++ .github/ISSUE_TEMPLATE/workflows/test_and_promote_model.yml
@@ -1,7 +1,11 @@
+# This does not work because nothing is hosted in a cloud so far
+# This is a potential cloud ci pipeline to test the model and promote it to staging if it passed all test
+
 name: Model Test & Staging Promotion
 
-on: [workflow_dispatch, push]
+on: [workflow_dispatch, push] # TODO WHAT IS WORKFLOW DISPATCH ???
 
+# TODO FINISH ... 
 jobs:
   test_and_stage:
     runs-on: ubuntu-latest
@@ -13,10 +17,8 @@ jobs:
           python-version: '3.11'
       - name: Install dependencies
         run: |
-          pip install -r api/requirements.txt
-          pip install -r shared/requirements.txt
-          pip install mlflow
+          pip install -r requirements.txt
       - name: Run model performance tests
-        run: python ci/test_model_performance.py
+        run: python ci/tests/pre_promotion_tests.py
       - name: Promote model to staging
-        run: python ci/promote_to_staging.py
+        run: python ci/promotion/promote_to_staging.py
