+++ api/model.py
@@ -3,16 +3,17 @@ import os
 import pandas as pd
 
 class Model:
-    def __init__(self, model_name: str, model_stage: str = "Production"):
+    def __init__(self, model_name):
         self.model_name = model_name
-        self.model_stage = model_stage
         self.model = None
 
     def load(self):
         if self.model is None:
             tracking_uri = os.environ.get("MLFLOW_URI", "http://mlflow:5000")
             mlflow.set_tracking_uri(tracking_uri)
-            model_uri = f"models:/{self.model_name}/{self.model_stage}"
+            # model_uri = f"models:/{self.model_name}/{self.model_stage}"
+            # TODO : LOAD MODEL BY ALIAS PRODUCTION
+            model_uri = f"models:/{self.model_name}@production"
             self.model = mlflow.pyfunc.load_model(model_uri)
 
     def predict(self, data):
