+++ ci/docker-compose.yml
@@ -1,0 +1,112 @@
+services:
+  db:
+    image: postgres:latest
+    restart: always
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: db
+      POSTGRES_HOST: db
+      POSTGRES_PORT: 5432
+    ports:
+      - 5432:5432
+    volumes:
+      - pg_data:/var/lib/postgresql/data
+    
+  db-init:
+    build:
+      context: .
+      dockerfile: db/Dockerfile
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: db
+      POSTGRES_HOST: db
+      POSTGRES_PORT: 5432
+      CRYPT_CONTEXT: bcrypt
+      TEST_USER: test_user
+      TEST_USER_PASSWORD: test_password
+    depends_on:
+      - db
+    restart: "no"
+
+  mlflow:
+    build:
+      context: .
+      dockerfile: mlflow/Dockerfile
+    image: ghcr.io/mlflow/mlflow
+    environment:
+      - MLFLOW_TRACKING_URI: postgresql://postgres:postgres@db:5432/db
+    command: >
+      mlflow server
+      --backend-store-uri postgresql://$postgres:$postgres@db:5432/$db
+      --default-artifact-root /mlflow
+      --host 0.0.0.0
+      --port 5000
+    ports:
+      - 5000:5000
+    volumes:
+      - mlruns:/mlflow
+    depends_on:
+      - db
+      - db-init
+    restart: always
+
+  training:
+    build:
+      context: .
+      dockerfile: training/Dockerfile
+    environment:
+      - POSTGRES_USER: postgres
+      - POSTGRES_PASSWORD: postgres
+      - POSTGRES_DB: db
+      - POSTGRES_HOST: db
+      - POSTGRES_PORT: 5432
+      - MLFLOW_URI: http://mlflow:5000
+    volumes:
+      - mlruns:/app/mlruns
+    depends_on:
+      - db
+      - db-init
+      - mlflow
+    restart: "no"
+
+  promotion:
+    build:
+      context: .
+      dockerfile: ci/promotion/Dockerfile
+    environment:
+      - MLFLOW_URI: http://mlflow:5000
+    volumes:
+      - mlruns:/app/mlruns
+    restart: "no"
+    
+  api:
+    build:
+      context: .
+      dockerfile: ./api/Dockerfile
+    restart: always
+    ports:
+      - ${API_LOCAL_PORT}:${API_PORT}
+    environment:
+      - POSTGRES_USER: postgres
+      - POSTGRES_PASSWORD: postgres
+      - POSTGRES_DB: db
+      - POSTGRES_HOST: db
+      - POSTGRES_PORT: 5432
+      - CRYPT_CONTEXT: bcrypt
+      - TEST_USER: test_user
+      - TEST_USER_PASSWORD: test_password
+      - MLFLOW_URI: http://mlflow:5000
+      - JWT_SECRET_KEY: test_key
+      - JWT_ALGORITHM: HS256
+      - ACCESS_TOKEN_EXPIRE_MINUTES: 30
+    depends_on:
+      - db
+      - db-init
+      - mlflow
+
+volumes:
+  pg_data:
+  mlruns:
+
